formatting and removal of unused variables (according to eclipse)
git-svn-id: https://svn.apache.org/repos/asf/incubator/wicket/trunk/wicket@457812 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/wicket/util/crypt/TrivialCrypt.java
@@ -42,7 +42,7 @@ public class TrivialCrypt extends AbstractCrypt
 		final byte[] result = new byte[input.length];
 		for (int i = 0; i < input.length; i++)
 		{
-			result[i] = (byte)((int)input[i] ^ 0xff);
+			result[i] = (byte)(input[i] ^ 0xff);
 		}
 		return result;
 	}
